Upgrade GHC to v9.6, and nixpkgs for good measure.
The modification to `tar` has been released, so we can use that version
now instead of a random commit.

Some types and functions in Aeson have been deprecated and moved.

We no longer need to patch Ormolu in nixpkgs, as the fix has been merged.

This also brings along an upgrade to Ruby, resulting in changes to error
message syntax, which means a change to a fixture.